+++ packages/botonic-nlu/src/index.js
@@ -1,0 +1,3 @@
+import { BotonicNLU } from './botonic-nlu'
+import * as CONSTANTS from './constants'
+export { BotonicNLU, CONSTANTS }
