+++ packages/botonic-nlu/src/preprocessing.js
@@ -1,7 +1,7 @@
 import * as tf from '@tensorflow/tfjs'
 import franc from 'franc'
-import { replaceAll, clone } from './utils'
-import { UNKNOWN_TOKEN } from './constants'
+import { replaceAll, clone, shuffle } from './utils'
+import { UNKNOWN_TOKEN, ENTITIES_REGEX } from './constants'
 
 export class Tokenizer {
   constructor(vocabulary = null) {
@@ -105,4 +105,56 @@ export function detectLang(input, langs) {
     res = langs[0]
   }
   return res
+}
+
+export function parseUtterance(utterance) {
+  let capturedGroup = utterance.match(new RegExp(ENTITIES_REGEX, 'g')) || []
+  let parsedEntities = capturedGroup
+    .map(matched => ENTITIES_REGEX.exec(matched))
+    .map(parsedEntity => ({
+      raw: parsedEntity[0],
+      value: parsedEntity[1],
+      type: parsedEntity[2]
+    }))
+  for (let entity of parsedEntities) {
+    utterance = utterance.replace(entity.raw, entity.value)
+  }
+  return { parsedUtterance: utterance, parsedEntities }
+}
+
+export function preprocessData(devIntents, params) {
+  let { samples, labels } = getSamplesAndLabels(devIntents.intents)
+  let tokenizer = new Tokenizer()
+  tokenizer.fitOnSamples(samples)
+  let sequences = tokenizer.samplesToSequences(samples)
+  let seqLength = params.MAX_SEQ_LENGTH || tokenizer.maxSeqLength
+  params.MAX_SEQ_LENGTH = seqLength
+  let tensorData = padSequences(sequences, seqLength)
+  console.log(`Shape of data tensor: [${tensorData.shape}]`)
+  let tensorLabels = tf.oneHot(
+    tf.tensor1d(labels, 'int32'),
+    Object.keys(devIntents.intentsDict).length
+  )
+  console.log(`Shape of label tensor: [${tensorLabels.shape}]`)
+  let vocabularyLength = tokenizer.vocabularyLength
+  console.log(`Found ${vocabularyLength} unique tokens`)
+  return {
+    tensorData,
+    tensorLabels,
+    vocabulary: tokenizer.vocabulary,
+    vocabularyLength: tokenizer.vocabularyLength
+  }
+}
+
+function getSamplesAndLabels(intents) {
+  let { samples, labels } = intents.reduce(
+    (result, intent) => {
+      result.samples.push(intent.utterance)
+      result.labels.push(intent.label)
+      return result
+    },
+    { samples: [], labels: [] }
+  )
+  shuffle(samples, labels)
+  return { samples, labels }
 }
