+++ packages/botonic-plugin-nlu/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@botonic/plugin-nlu",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "main": "src/index.js",
   "scripts": {
     "build": "rm -rf dist && babel src -d dist",
@@ -9,39 +9,18 @@
   "jest": {
     "testEnvironment": "node"
   },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "current"
-          }
-        }
-      ]
-    ]
-  },
   "dependencies": {
-    "@tensorflow/tfjs": "^1.2.7",
-    "@tensorflow/tfjs-node": "^1.2.7",
-    "@types/jest": "^24.0.17",
+    "@babel/runtime": "^7.5.5",
+    "@botonic/nlu": "0.1.0",
+    "@tensorflow/tfjs": "^1.2.8",
     "axios": "latest",
-    "colors": "^1.3.3",
-    "compromise": "^11.13.2",
-    "compromise-plugin": "0.0.9",
-    "franc": "^4.1.0",
-    "fs": "0.0.1-security",
-    "inquirer": "^6.3.1",
-    "jest": "^24.8.0",
-    "node-fetch": "^2.6.0",
-    "node-nlp": "^3.5.2",
-    "sqlite-async": "^1.0.11"
+    "node-fetch": "^2.6.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
     "@babel/core": "^7.5.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.5.0",
     "@babel/preset-env": "^7.5.5",
-    "@babel/runtime": "^7.5.5",
     "core-js": "^3.1.4"
   }
 }
